+++ src/Contracts/DialEventArgs.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace DialToolsForVS
+{
+    public class DialEventArgs : EventArgs
+    {
+        public bool Handled { get; set; }
+        public string Action { get; set; }
+    }
+}