+++ src/Controller/StatusBarControl.cs
@@ -19,7 +19,7 @@ namespace DialToolsForVS
             {
                 Width = 12,
                 Height = 12,
-                Margin = new Thickness(5, 4, 0, 5),
+                Margin = new Thickness(5, 1, 3, 0),
                 Stroke = Brushes.WhiteSmoke,
                 StrokeThickness = 2
             };