+++ src/Providers/SolutionExplorer/SolutionExplorerController.cs
@@ -19,12 +19,12 @@ namespace DialToolsForVS
 
         public bool CanHandleRotate => false;
 
-        public bool OnClick(RadialControllerButtonClickedEventArgs args)
+        public void OnClick(RadialControllerButtonClickedEventArgs args, DialEventArgs e)
         {
             var selectedItems = VsHelpers.DTE.ToolWindows.SolutionExplorer.SelectedItems as UIHierarchyItem[];
 
             if (selectedItems == null || selectedItems.Length != 1)
-                return false;
+                return;
 
             if (selectedItems[0].UIHierarchyItems.Expanded)
             {
@@ -35,12 +35,11 @@ namespace DialToolsForVS
                 SendKeys.Send("{RIGHT}");
             }
 
-            return true;
+            e.Handled = true;
         }
 
-        public bool OnRotate(RotationDirection direction)
+        public void OnRotate(RotationDirection direction, DialEventArgs e)
         {
-            return false;
         }
     }
 }