+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Road map
 
 - [x] Status bar text update on activation state change
-- [ ] Write dial event description in status bar
+- [x] Write dial event description in status bar
+- [ ] Output Window logging
 - [ ] Add telemtry
 
 Features that have a checkmark are complete and available for